Getting rid of jest for now
